fix: using price int for cosmetics

+++ Runtime/Models/Creatables/CosmeticCategoryEntry.cs
@@ -10,11 +10,11 @@ namespace FTRShared.Runtime.Models
 
         public string url_id = "";
 
-        public float price = 1f;
+        public int price = 1;
 
         public CosmeticCategoryEntry() { }
 
-        public CosmeticCategoryEntry(string spritePath, string urlId = "", float price = 1f)
+        public CosmeticCategoryEntry(string spritePath, string urlId = "", int price = 1)
         {
             this.sprite_path = spritePath ?? "";
             this.url_id = urlId ?? "";
